refactor: move @file extraction logic into separate file
This will make it easier to test, and also allow adding more sophisticated
extraction logic without bloating `session/index.ts`.
